requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3

gitpython==3.1.43

faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
tqdm==4.66.4

google-generativeai==0.7.2

tree_sitter==0.21.1
tree_sitter_languages==1.10.2

#database dependencies

sqlalchemy
psycopg2-binary
pgvector

